api/archive/source/test-connection/requirements.txt
pymysql
pymssql
oracledb
psycopg2-binary==2.9.6